Refactorizar Idioma Ahora con llamar al archivo idioma.js se carga todo.
