+++ editor.html
@@ -85,6 +85,7 @@
 	</main>
 
 	<script src="funciones/configuración/tema.js"></script>
+	<script src="funciones/configuración/idioma.js"></script>
 	<script type="module" src="funciones/creación.js"></script>
 	<script type="module" src="funciones/visualizar.js"></script>
 	<script type="module" src="funciones/principal.js"></script>
