+++ funciones/configuración/idioma.js
@@ -1,16 +1,44 @@
-// ¡Refactorizar e implementar los idiomas!
-export default function cambiar_idioma(idioma) {
-	const lang = {
-		"deutsch": "de", "english": "en", "español": "es", "français": "fr", "italiano": "it", "português": "pt", "中文": "zh", "日本語": "ja", "한국어": "ko"
-	};
-	const banderas = {
-		"español": "es", "english": "uk", "français": "fr", "deutsch": "de", "italiano": "it", "português": "pt", "中文": "cn", "日本語": "jp", "한국어": "kr"
-	};
+const botón_idioma = document.getElementById("botón_idioma");
+const listado_idioma = document.getElementById("lista_idioma");
 
-	// Poner la bandera del idioma en el botón del cambio de idioma.
-	const nodo_bandera = document.getElementById("botón_idioma");
-	nodo_bandera.setAttribute("src", "archivos/imágenes/banderas/" + banderas[idioma] + ".png");
+const idiomas = {
+	"español": "es", "english": "en", "français": "fr", "deutsch": "de", "italiano": "it", "português": "pt", "中文": "zh", "日本語": "ja", "한국어": "ko"
+};
 
-	// Ponemos la página web en el idioma correspondiente.
-	document.querySelector("html").setAttribute("lang", lang[idioma]);
+const idiomas_disponibles = ["español", "english"];
+
+crear_listado_idiomas();
+
+botón_idioma.addEventListener("click", () => {
+	listado_idioma.classList.toggle("oculto");
+});
+
+function cambiar_idioma(idioma) {
+	botón_idioma.setAttribute("src", "archivos/imágenes/banderas/" + idiomas[idioma] + ".png");
+	document.querySelector("html").setAttribute("lang", idiomas[idioma]);
+	listado_idioma.classList.toggle("oculto");
+}
+
+function crear_listado_idiomas() {
+	const fragmento = document.createDocumentFragment();
+
+	for (const idioma of idiomas_disponibles) {
+		const elemento = document.createElement("li");
+
+		const imagen = document.createElement("img");
+		imagen.setAttribute("src", "archivos/imágenes/banderas/" + idiomas[idioma] + ".png");
+		elemento.appendChild(imagen);
+
+		const texto = document.createElement("span");
+		texto.textContent = " " + idioma.charAt(0).toUpperCase() + idioma.substring(1);
+		elemento.appendChild(texto);
+
+		elemento.addEventListener("click", () => {
+			cambiar_idioma(idioma);
+		});
+
+		fragmento.appendChild(elemento);
+	}
+
+	listado_idioma.appendChild(fragmento);
 }
