+++ index.html
@@ -37,6 +37,7 @@
 	</main>
 
 	<script src="funciones/configuración/tema.js"></script>
+	<script src="funciones/configuración/idioma.js"></script>
 	<script src="funciones/colección.js"></script>
 	<script type="module" src="funciones/principal.js"></script>
 	<script src="funciones/almacenamiento.js"></script>
